refactor: crawler와 openapi 패키지를 global 패키지로 이동
