+++ src/main/java/com/bookspot/batch/global/crawler/common/CookieKeyConst.java
@@ -1,4 +1,4 @@
-package com.bookspot.batch.crawler.common;
+package com.bookspot.batch.global.crawler.common;
 
 public class CookieKeyConst {
     public static final String SESSION_ID = "JSESSIONID";
