+++ src/main/java/com/bookspot/batch/global/openapi/naru/NaruApiUrlCreator.java
@@ -1,4 +1,4 @@
-package com.bookspot.batch.openapi.naru;
+package com.bookspot.batch.global.openapi.naru;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
